+++ tests/test_tibberlib_spec.js
@@ -1,8 +1,10 @@
 /* eslint-env mocha */
 'use strict';
 var should = require('should');
+var WebSocket = require('ws');
 var tibberLib = require('../nodes/tibbermodule.js');
 var testConfig = 'envt-test';
+var testSub = { query: 'subscription{ liveMeasurement(homeId:"c70dcbe5-4485-4821-933d-a8a86452737b"){timestamp power maxPower accumulatedConsumption accumulatedCost}}' };
 
 describe('Tibberlib module function tests', function () {
   // Tibber module should be able to fetch token. Future release with oauth client, currently using pre-reqistereded manual token from file
@@ -62,6 +64,28 @@ describe('Tibberlib module function tests', function () {
       price.homes[0].currentSubscription.priceInfo.current.should.have.property('total');
       price.homes[0].currentSubscription.priceInfo.current.should.have.property('tax');
       done();
+    }).catch((err) => {
+      console.log('error: ' + err);
+      console.dir(err);
+    });
+  });
+
+  // Tibber module should be able to subscribe and get a ws
+  it('should subscribe', function (done) {
+    tibberLib.readConfig(testConfig);
+    tibberLib.subscribe(testSub).then((result, err) => {
+      let ws = result;
+      if(typeof ws.error !== 'undefined' && ws.error) {console.dir(ws);}
+      ws.on('message', function (data) {
+        console.log(data);
+      });
+      ws.on('error', function (data) {
+        console.log('open error! ' + data);
+        done();
+      });
+      console.dir(ws);
+      ws.should.be.an.instanceOf(WebSocket);
+      // done();
     }).catch((err) => {
       console.log('error: ' + err);
       console.dir(err);